% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_digit1.R
\name{detect_digit1}
\alias{detect_digit1}
\title{Detect digit that should be read as 'Aho'}
\usage{
detect_digit1(x, d = 3, engine = c("base", "stringi", "stringzilla"))
}
\arguments{
\item{x}{Integers.}

\item{d}{Integer scalar; a digit to be detected.}

\item{engine}{One of 'base', 'stringr', or 'stringzilla'.}
}
\value{
Logicals.
}
\description{
Detect whether the integers are multiples of \code{d} or
numbers where \code{d} is used as one of the digits.
}
