% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nebeatsu.R
\docType{package}
\name{nabeatsuR-package}
\alias{nabeatsuR}
\alias{nabeatsuR-package}
\title{nabeatsuR: 'Nabeatsu' of the World R Package}
\description{
Detects numbers that you should say as 'Aho'.
}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

Authors:
\itemize{
  \item Ash Vardanian [copyright holder]
}

}
