% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nebeatsu.R
\name{should_be_aho}
\alias{should_be_aho}
\title{Detect digit that should be said as 'Aho'}
\usage{
should_be_aho(x, d = 3, engine = NULL)
}
\arguments{
\item{x}{Integers.}

\item{d}{Integer scalar; a digit to be detected.}

\item{engine}{One of 'base', 'stringr', or 'stringzilla'.}
}
\value{
Logicals.
}
\description{
This function is for '3' only.
The arguments \code{d} and \code{engine} are not used in this function.
}
