% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strzl_count.R
\name{strzl_count}
\alias{strzl_count}
\title{Count number of matches}
\usage{
strzl_count(string, pattern, .na = FALSE)
}
\arguments{
\item{string}{Character vector.}

\item{pattern}{String scalar. Unlike \code{stringr::str_count},
\code{pattern} must be a single string.　The second and subsequent strings are ignored.}

\item{.na}{If \code{TRUE}, returns \code{NA} for an \code{NA} string.}
}
\value{
Integers.
}
\description{
Counts the number of times \code{pattern} is found
within each element of \code{string}.
}
