% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strzl_detect.R
\name{strzl_detect}
\alias{strzl_detect}
\title{Detect the presence of a match}
\usage{
strzl_detect(string, pattern, .na = FALSE)
}
\arguments{
\item{string}{Character vector.}

\item{pattern}{String scalar; pattern to look for.}

\item{.na}{If \code{TRUE}, returns \code{NA} for an \code{NA} string.}
}
\value{
Logicals.
}
\description{
Equivalent to \code{gerpl(pattern, string)}.
}
