% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strzl_which.R
\name{strzl_which}
\alias{strzl_which}
\title{Find matcing indices}
\usage{
strzl_which(string, pattern)
}
\arguments{
\item{string}{Character vector.}

\item{pattern}{String scalar; pattern to look for.}
}
\value{
Integers.
}
\description{
Equvalent to \code{grep(pattern, string)}.
}
